+++ lib/slick_core.dart
@@ -5,6 +5,12 @@ import 'dart:html' as html;
 //import 'dart:math' as math;
 //import 'dart:mirrors';
 EditorLock GlobalEditorLock = new EditorLock();
+
+//parm:  List<core.Range>, Map
+//e: KeyboardEvent, EventData
+typedef handlerFunction(Object e,Object parm);
+
+
 /**
  * utility to get dom width / height
  */
@@ -89,7 +95,7 @@ class Event {
    * @method subscribe
    * @param fn {Function} Event handler.
    */
-  subscribe (Function fn) {
+  subscribe (handlerFunction fn) {
     handlers.add(fn);
   }
 
@@ -114,7 +120,7 @@ class Event {
    *      If not specified, the scope will be set to the <code>Event</code> instance.
    */
   notify(args, [dynamic e, scope]) {
-    if(e==null) e = new EventData();
+    e ??= new EventData();
    // scope = scope || this;
     var returnValue;
     for (int i = 0; i < handlers.length && !(e is EventData && (e.isPropagationStopped() || e.isImmediatePropagationStopped())); i++) {
@@ -280,7 +286,7 @@ class NonDataItem {
  * @constructor
  */
 class Group extends NonDataItem {
-  bool __group = true;
+ // bool __group = true;
 
   /**
    * Grouping level, starting with 0.
@@ -374,7 +380,7 @@ class Group extends NonDataItem {
  * @constructor
  */
 class GroupTotals extends NonDataItem{
-  bool __groupTotals = true;
+//  bool __groupTotals = true;
 
   /***
    * Parent Group.